code/requirements.txt
psycopg2-binary>=2.8.3
pika>=1.1.0
